# feat: return images from resources when sync occurs