--- a/pkg/sync/common/types.go
+++ b/pkg/sync/common/types.go
@@ -136,7 +136,9 @@ func NewHookDeletePolicy(p string) (HookDeletePolicy, bool) {
 type ResourceSyncResult struct {
 	// holds associated resource key
 	ResourceKey kube.ResourceKey
-	// holds the images associated with the resource
+	// Images holds the images associated with the resource. These images are collected on a best-effort basis
+	// from fields used by known workload resources. This does not necessarily reflect the exact list of images
+	// used by workloads in the application.
 	Images []string
 	// holds resource version
 	Version string