--- a/pkg/utils/kube/kube_test.go
+++ b/pkg/utils/kube/kube_test.go
@@ -208,7 +208,7 @@ spec:
 	require.NoError(t, err)
 
 	images := GetResourceImages(&deployment)
-	assert.Equal(t, expected, images)
+	require.Equal(t, expected, images)
 }
 
 func TestGetResourceImagesForPod(t *testing.T) {
@@ -236,7 +236,7 @@ spec:
 	require.NoError(t, err)
 
 	images := GetResourceImages(&pod)
-	assert.Equal(t, expected, images)
+	require.Equal(t, expected, images)
 }
 
 func TestSplitYAML_SingleObject(t *testing.T) {