--- a/pkg/utils/kube/kube_test.go
+++ b/pkg/utils/kube/kube_test.go
@@ -239,7 +239,7 @@ spec:
 	require.Equal(t, expected, images)
 }
 
-func TestGetImages_NoImagesPresent(t *testing.T) {
+func TestGetImagesNoImagesPresent(t *testing.T) {
 	manifests := [][]byte{
 		[]byte(`
 apiVersion: v1