--- a/pkg/utils/kube/kube_test.go
+++ b/pkg/utils/kube/kube_test.go
@@ -179,13 +179,11 @@ spec:
 }
 
 func TestGetResourceImages(t *testing.T) {
-	type testCase struct {
+	testCases := []struct {
 		manifest    []byte
 		expected    []string
 		description string
-	}
-
-	testCases := []testCase{
+	}{
 		{
 			manifest: []byte(`
 apiVersion: extensions/v1beta2