--- a/pkg/utils/kube/kube_test.go
+++ b/pkg/utils/kube/kube_test.go
@@ -250,22 +250,8 @@ spec:
 			expected:    []string{"busybox:1.28"},
 			description: "cronjob with containers",
 		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.description, func(t *testing.T) {
-			resource := unstructured.Unstructured{}
-			err := yaml.Unmarshal(tc.manifest, &resource)
-			require.NoError(t, err)
-			images := GetResourceImages(&resource)
-			require.Equal(t, tc.expected, images)
-		})
-	}
-}
-
-func TestGetImagesNoImagesPresent(t *testing.T) {
-	manifests := [][]byte{
-		[]byte(`
+		{
+			manifest: []byte(`
 apiVersion: v1
 kind: ConfigMap
 metadata:
@@ -280,7 +266,11 @@ data:
     key3=value3
   log.level: debug
 `),
-		[]byte(`
+			expected:    nil,
+			description: "configmap without containers",
+		},
+		{
+			manifest: []byte(`
 apiVersion: apps/v1
 kind: Deployment
 metadata:
@@ -302,7 +292,11 @@ spec:
         configMap:
           name: config
 `),
-		[]byte(`
+			expected:    nil,
+			description: "deployment without containers",
+		},
+		{
+			manifest: []byte(`
 apiVersion: apps/v1
 kind: Deployment
 metadata:
@@ -317,7 +311,11 @@ spec:
         - name: text-service
           command: ["echo", "hello"]
 `),
-		[]byte(`
+			expected:    nil,
+			description: "deployment with container without image",
+		},
+		{
+			manifest: []byte(`
 apiVersion: v1
 kind: Pod
 metadata:
@@ -328,17 +326,20 @@ spec:
   containers:
   - name: no-image-container
     command: ["echo", "hello"]
-`,
-		),
+`),
+			expected:    nil,
+			description: "pod with container without image",
+		},
 	}
 
-	for _, manifest := range manifests {
-		resource := unstructured.Unstructured{}
-		err := yaml.Unmarshal([]byte(manifest), &resource)
-		require.NoError(t, err)
-
-		images := GetResourceImages(&resource)
-		require.Empty(t, images)
+	for _, tc := range testCases {
+		t.Run(tc.description, func(t *testing.T) {
+			resource := unstructured.Unstructured{}
+			err := yaml.Unmarshal(tc.manifest, &resource)
+			require.NoError(t, err)
+			images := GetResourceImages(&resource)
+			require.Equal(t, tc.expected, images)
+		})
 	}
 }