Implementing the Gitops side of this issue: https://github.com/argoproj/argo-cd/issues/20896

Adds an api to `kube.go` to get the images from a resource. Adds those images to `ResourceSyncResult`. This API accounts for how different k8s resources define the images.