@rumstead 2025-02-18
Can you create a PR to Argo CD updating the `go.mod` to your branch? That will allow us to validate the e2e tests run and pass.