@Aaron-9900 2025-02-20
@rumstead done in https://github.com/argoproj/argo-cd/pull/20954, tests are now passing